# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Financial Analysis Pipeline — REST API",
    "version": "1.0.0",
    "description": "REST surface of the Financial Analysis Pipeline analysis server (Express + Socket.IO on :3001). Covers connection config, LLM provider config, per-analyst source credentials / weights / flavors, market data (quote, history, options history, company news + sentiment), the Stock Screener, the multi-agent registry (agencies + analysts CRUD), and report persistence. Realtime analysis streaming itself is delivered over Socket.IO (out of scope for this REST doc).",
    "contact": { "name": "Financial Analysis Pipeline" }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev server" }
  ],
...
  "paths": {
...
    "/analyst-params": {
      "get": {
        "tags": ["analyst-params"],
        "summary": "Saved weights for an agency",
        "parameters": [
          { "name": "sessionId", "in": "query", "schema": { "type": "string" } },
          { "name": "agencyId", "in": "query", "schema": { "type": "string", "default": "long-term" } }
        ],
        "responses": {
          "200": { "description": "Params", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalystParamsGet" } } } },
          "404": { "description": "Unknown agency", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["analyst-params"],
        "summary": "Store one analyst's weight overrides",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalystParamsPost" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Analyst not in agency", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/analyst-flavors": {
      "get": {
        "tags": ["analyst-flavors"],
        "summary": "Resolved flavor set for one analyst",
        "parameters": [
          { "name": "sessionId", "in": "query", "schema": { "type": "string" } },
          { "name": "agencyId", "in": "query", "schema": { "type": "string", "default": "long-term" } },
          { "name": "analystId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Flavors", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalystFlavorsGet" } } } },
          "404": { "description": "Unknown agency", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["analyst-flavors"],
        "summary": "Replace a user's flavor set + selection",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AnalystFlavorsPost" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" } } } } } },
          "400": { "description": "Invalid flavor set", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/analyst-flavors/agency-summary": {
      "get": {
        "tags": ["analyst-flavors"],
        "summary": "LLM enabled/disabled summary for an agency",
        "parameters": [
          { "name": "sessionId", "in": "query", "schema": { "type": "string" } },
          { "name": "agencyId", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Summary", "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
    "/analyst-flavors/bulk-enable-llm": {
      "post": {
        "tags": ["analyst-flavors"],
        "summary": "Enable/disable LLM for all analysts in an agency",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "sessionId": { "type": "string" }, "agencyId": { "type": "string" }, "enabled": { "type": "boolean" } } } } } },
        "responses": {
          "200": { "description": "Done", "content": { "application/json": { "schema": { "type": "object" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": { "type": "object", "properties": { "error": { "type": "string" }, "details": { "type": "array", "items": { "type": "string" } } } },
...
      "AnalystParamsGet": { "type": "object", "properties": { "sessionId": { "type": "string" }, "agencyId": { "type": "string" }, "params": { "type": "object", "additionalProperties": { "type": "object", "additionalProperties": { "type": "number" } } } } },
      "AnalystParamsPost": { "type": "object", "properties": { "sessionId": { "type": "string" }, "agencyId": { "type": "string" }, "analystId": { "type": "string" }, "params": { "type": "object", "additionalProperties": { "type": "number" } } } },
      "AnalystFlavorsGet": { "type": "object", "properties": { "flavors": { "type": "array", "items": { "type": "object" } }, "selectedId": { "type": "string", "nullable": true } } },
      "AnalystFlavorsPost": { "type": "object", "properties": { "sessionId": { "type": "string" }, "agencyId": { "type": "string" }, "analystId": { "type": "string" }, "flavors": { "type": "array", "items": { "type": "object" } }, "selectedId": { "type": "string" } } },
...
    }
  }
}